+++ compiler/testData/diagnostics/testsWithStdLib/annotations/strictfpOnClass.kt
@@ -1,11 +1,17 @@
-<!WRONG_ANNOTATION_TARGET!>@kotlin.jvm.Strictfp<!> class A {
+<!STRICTFP_ON_CLASS!>@kotlin.jvm.Strictfp<!> class A {
 
 }
 
-<!WRONG_ANNOTATION_TARGET!>@kotlin.jvm.Strictfp<!> object B {
+<!STRICTFP_ON_CLASS!>@kotlin.jvm.Strictfp<!> object B {
 
 }
 
-<!WRONG_ANNOTATION_TARGET!>@kotlin.jvm.Strictfp<!> interface C {
+<!STRICTFP_ON_CLASS!>@kotlin.jvm.Strictfp<!> interface C {
 
+}
+
+fun foo() {
+    <!STRICTFP_ON_CLASS!>@kotlin.jvm.Strictfp<!> class D
+
+    <!STRICTFP_ON_CLASS!>@kotlin.jvm.Strictfp<!> object: Any() {}
 }